+++ eslint.config.mjs
@@ -1,7 +1,7 @@
 import { generateEslintConfig } from '@companion-module/tools/eslint/config.mjs'
 
 const baseConfig = await generateEslintConfig({
-	ignores: ['pkg/**', 'node_modules/**', 'doc/**'],
+	ignores: ['pkg/**', 'node_modules/**', 'doc/**', 'tools/**'],
 })
 
 export default [